refactor: restructure Rewind import per PR review
Addresses review feedback on #2846:

- Move rewind import from screenpipe-engine to screenpipe-connect
  as it's an integration/connection, not engine functionality
- Split 872-line monolith into scanner.rs, video_importer.rs,
  audio_importer.rs, checkpoint.rs + orchestration in mod.rs
- Replace direct whisper-rs usage with POST to the existing
  /v1/audio/transcriptions endpoint, removing the whisper-rs
  dependency from both screenpipe-engine and screenpipe-app
  and respecting the user's configured transcription engine
- Reuse a single DatabaseManager in RewindImportState instead
  of creating a fresh one per Tauri command
- Accept OCR as a callback (OcrFn) to avoid a cyclic dependency
  with screenpipe-screen; callers inject perform_ocr_apple
- Add 17 unit tests covering date filtering, checkpoint round-trip,
  image hashing, and state serialization
