+++ apps/screenpipe-app-tauri/src-tauri/Cargo.toml
@@ -127,7 +127,6 @@ screenpipe-engine = { path = "../../../crates/screenpipe-engine", default-featur
 screenpipe-screen = { path = "../../../crates/screenpipe-screen" }
 screenpipe-a11y = { path = "../../../crates/screenpipe-a11y" }
 screenpipe-audio = { path = "../../../crates/screenpipe-audio" }
-whisper-rs = { git = "https://github.com/screenpipe/whisper-rs.git", rev = "97e4070", features = ["tracing_backend"] }
 screenpipe-db = { path = "../../../crates/screenpipe-db" }
 screenpipe-vault = { path = "../../../crates/screenpipe-vault" }
 screenpipe-connect = { path = "../../../crates/screenpipe-connect" }
